Add files via upload
Hi I have added in the database searching pipeline I have used in my application. 

Here is a short description of it. 

Essentially there are two parts. 
1)	Db cleaning pipeline
2)	Db searching pipeline

The db cleaning pipeline uses the CompoundDB package to extract a ‘normalised’ db from sdf files (this is a pretty standard format for most major db, also the Fiehn lab have made their own downloadable versions, here https://massbank.us/downloads). 
I then clean up the databases that are extracted by removing any NA values and zero values etc. This is followed by calculating adducts and appending them on to the db dataframe. 

Next I then use the clean databases in the searching pipepline, I’ve described the three steps of the pipeline in the source code, But what it does in essence is filters the db for adducts of interest and elements of interest. Then the searching algorithm goes through and matches your m/z values against those in the db using a ppm range set by the user (adducts are also set by the user). 

I have written the pipeline in a way where you could feasible put in your own custom DB. If you need help with developing a custom DB please let me know. :)

A couple of notes on the db structure that I use. In the first step of cleaning the DB’s I summarise the DB in terms of isomers, i.e., matching formulas and exact masses. This means that you can have a 1-1, 1-many when matching your mz  value. However, all the possible isomers for a given combination of formula and exact mass are given as a single match, hopefully simplifying the process. Again, just want to say that this is a level 4 match at best when considering the definition by the International Metabolomics Standards Initiative. 

I have also added in a script called DBsearcher_pipeline_custom.R.
This does retention time searching (again not relevant here, but this could easily be adapted to another orthogonal separation technique like ion mobility CCS values)

Another note of difference with my pipeline. I noticed that you guys only use an absolute difference of 0.005 Da to match masses. This in general is OK. I use ppm, this makes it specific to each molecule that you are matching. Using ppm will give more certainty in matching to databases.